parse IR

+++ core/expr/Cast/SExtExpr.cpp
@@ -1,6 +1,6 @@
 #include "SExtExpr.h"
 
-SExtExpr::SExtExpr(const std::shared_ptr<Expr>& base, unsigned int width):Expr(width,0)
+SExtExpr::SExtExpr(const std::shared_ptr<Expr>& base, uint64_t width):Expr(width,0)
 {
 	rightExpr = base;
 }
