+++ core/expr/Cast/ZExtExpr.cpp
@@ -1,6 +1,6 @@
 #include "ZExtExpr.h"
 
-ZExtExpr::ZExtExpr(const std::shared_ptr<Expr>& base, unsigned int width) :Expr(width, 0)
+ZExtExpr::ZExtExpr(const std::shared_ptr<Expr>& base, uint64_t width) :Expr(width, 0)
 {
 }
